simplified the logic as expiry date generation is only happening in case of annual test types
