--- src/models/IMaxDates.ts
@@ -1,4 +1,0 @@
-export interface IMaxDates {
-    maxDateForAllTestTypes: Date;
-    maxDateForAnnualTestTypes: Date;
-}
